python/requirements.txt
black
msoffcrypto-tool
openpyxl
